Split out getPlanDiscountedRawPrice
