+++ client/state/sites/plans/selectors/index.js
@@ -19,30 +19,7 @@ export { getPlansBySite, getPlansBySiteId } from 'state/sites/plans/selectors/ge
 export { getCurrentPlan } from 'state/sites/plans/selectors/get-current-plan';
 export { getSitePlan } from 'state/sites/plans/selectors/get-site-plan';
 export { isSitePlanDiscounted } from 'state/sites/plans/selectors/is-site-plan-discounted';
-
-/**
- * Returns a plan price, including any applied discounts
- *
- * @param  {object}  state         global state
- * @param  {number}  siteId        the site id
- * @param  {string}  productSlug   the plan product slug
- * @param  {boolean} isMonthly     if true, returns monthly price
- * @returns {number}                plan discounted raw price
- */
-export function getPlanDiscountedRawPrice(
-	state,
-	siteId,
-	productSlug,
-	{ isMonthly = false } = {}
-) {
-	const plan = getSitePlan( state, siteId, productSlug );
-
-	if ( get( plan, 'rawPrice', -1 ) < 0 || ! isSitePlanDiscounted( state, siteId, productSlug ) ) {
-		return null;
-	}
-	const discountPrice = plan.rawPrice;
-	return isMonthly ? calculateMonthlyPriceForPlan( productSlug, discountPrice ) : discountPrice;
-}
+export { getPlanDiscountedRawPrice } from 'state/sites/plans/selectors/get-plan-discounted-raw-price';
 
 /**
  * Returns a plan price before discount
